backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM Libraries
groq==0.4.2
anthropic==0.18.1
google-generativeai==0.3.2

# Database
motor==3.3.2
pymongo[srv,tls]==4.6.1
dnspython==2.7.0
certifi==2024.8.30

# Testing
httpx==0.25.2 